requirements.txt
PyGithub==1.57
PyYAML==6.0
tqdm==4.64.1
Jinja2==3.1.2
